@@ -2,8 +2,8 @@
 
 namespace Intervention\Gif\Test;
 
-use Intervention\Gif\Color;
-use Intervention\Gif\ColorTable;
+use Intervention\Gif\Blocks\Color;
+use Intervention\Gif\Blocks\ColorTable;
 
 class ColorTableTest extends BaseTestCase
 {
@@ -28,7 +28,7 @@ class ColorTableTest extends BaseTestCase
         $this->assertIsArray($table->getColors());
         $this->assertCount(2, $table->getColors());
 
-        foreach ($table->getColors() as $key => $value) {
+        foreach (array_keys($table->getColors()) as $key) {
             $this->assertIsNumeric($key);
         }
     }
@@ -119,9 +119,7 @@ class ColorTableTest extends BaseTestCase
     public function testDecode()
     {
         $source = "\x00\x00\x00\xff\x00\x00\xff\xff\x00\xff\xff\xff";
-        $table = ColorTable::decode($this->getTestHandle($source), function ($decoder) {
-            $decoder->setLength(12);
-        });
+        $table = ColorTable::decode($this->getTestHandle($source), 12);
 
         $this->assertInstanceOf(ColorTable::class, $table);
         $this->assertEquals(4, $table->countColors());
